Rename mod to gen9ssb
