+++ data/mods/gen9ssb/abilities.ts
