+++ data/mods/gen9ssb/conditions.ts
