+++ data/mods/gen9ssb/items.ts
