+++ data/mods/gen9ssb/pokedex.ts
