+++ data/mods/gen9ssb/random-teams.ts
