+++ data/mods/gen9ssb/scripts.ts
