+++ server/chat-plugins/randombattles/index.ts
@@ -6,7 +6,7 @@
  */
 
 import {FS, Utils} from '../../../lib';
-import {SSBSet, ssbSets} from '../../../data/mods/ssb/random-teams';
+import {SSBSet, ssbSets} from '../../../data/mods/gen9ssb/random-teams';
 
 
 interface SetCriteria {
